Grouped Calculations

mutate_groups(): Group the data, .data, using dplyr::group_by() then apply calculations using dplyr::mutate().

summarise_groups(): Group the data, .data, using dplyr::group_by() then summarise it by applying a calculations using dplyr::summarise().

mutate_groups(.data, groups, calculations)

summarise_groups(.data, groups, calculations)

Arguments

.data
A Spark DataFrame or a data.frame. The data to mutate or summarise.

groups
character(n). The columns to group by.

calculations
formula() or a list() of formula()s. The calculations to apply where the LHS of the formula is the name of the new column to create and the RHS is the calculation to apply.

Value

A Spark DataFrame or a data.frame depending on the input, .data.

Details

These functions offer the benefit over the scoped variants of being able to explicitly specify the parameters for each expression to evaluate.

Examples

mtcars %>%
  mutate_groups(
    groups = c("am", "cyl"),
    calculations = avgMpg ~ mean(mpg)
  )

#> # A tibble: 32 x 12
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb avgMpg
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4   20.6
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4   20.6
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1   28.1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1   19.1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2   15.0
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1   19.1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4   15.0
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2   22.9
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2   22.9
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4   19.1
#> # … with 22 more rows

mtcars %>%
  summarise_groups(
    groups = c("am", "cyl"),
    calculations = avgMpg ~ mean(mpg)
  )

#> # A tibble: 6 x 3
#>      am   cyl avgMpg
#>   <dbl> <dbl>  <dbl>
#> 1     0     4   22.9
#> 2     0     6   19.1
#> 3     0     8   15.0
#> 4     1     4   28.1
#> 5     1     6   20.6
#> 6     1     8   15.4

# You can pass lists of formulas
mtcars %>%
  mutate_groups(
    groups = c("am", "cyl"),
    calculations = list(
      avgMpg ~ mean(mpg, na.rm = TRUE),
      avgDisp ~ mean(disp, na.rm = TRUE)
    )
  )

#> # A tibble: 32 x 13
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb avgMpg
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4   20.6
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4   20.6
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1   28.1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1   19.1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2   15.0
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1   19.1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4   15.0
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2   22.9
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2   22.9
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4   19.1
#> # … with 22 more rows, and 1 more variable: avgDisp <dbl>